feat: twitter header download (#2594)
