+++ packages/shared/src/components/profile/devcard/index.ts
@@ -1,1 +1,7 @@
 export * from './DevCard';
+export * from './DevCardTwitterCover';
+export * from './DevCardStats';
+export * from './DevCardStatsSection';
+export * from './DevCardFooter';
+export * from './DevCardContainer';
+export * from './common';
