+++ packages/shared/src/graphql/users.ts
@@ -6,7 +6,7 @@ import {
 } from './fragments';
 import type { PublicProfile } from '../lib/user';
 import { Connection } from './common';
-import { Source, SourceMember } from './sources';
+import { SourceMember } from './sources';
 import { graphqlUrl } from '../lib/config';
 
 type PostStats = {
@@ -473,19 +473,6 @@ export const getReadingStreak = async (): Promise<UserStreak> => {
   return res.userStreak;
 };
 
-export interface DevCardData {
-  id: string;
-  user: PublicProfile;
-  createdAt: string;
-  theme: string;
-  isProfileCover: boolean;
-  showBorder: boolean;
-  reputation: number;
-  articlesRead: number;
-  tags: string[];
-  sources: Source[];
-}
-
 export const DEV_CARD_QUERY = gql`
   query DevCardById($id: ID!) {
     devCard(id: $id) {
@@ -497,10 +484,13 @@ export const DEV_CARD_QUERY = gql`
       theme
       isProfileCover
       showBorder
-      reputation
       articlesRead
       tags
-      sources
+      sources {
+        name
+        permalink
+        image
+      }
     }
   }
   ${USER_SHORT_INFO_FRAGMENT}
